change type of the id (long -> int) in all entities & update registerNewTicket in service layer

+++ src/main/java/ir/maktab58/BusTicketBooking/dao/BookingInfoDao.java
@@ -198,7 +198,7 @@ public class BookingInfoDao extends BaseDaoImpl<BookingInfo> {
         return list;
     }
 
-    public List getListOfFilteredBookingInfoByPriceRange(String source, String destination, int offset, long upBound, long downBound) {
+    public List getListOfFilteredBookingInfoByPriceRange(String source, String destination, int offset, int upBound, int downBound) {
         Session session = SessionUtil.getSession();
         Transaction transaction = session.beginTransaction();
 
@@ -229,7 +229,7 @@ public class BookingInfoDao extends BaseDaoImpl<BookingInfo> {
         return list;
     }
 
-    public List getListOfFilteredBookingInfoByPriceRange(String source, String destination, int offset, long upBound, long downBound, Date date) {
+    public List getListOfFilteredBookingInfoByPriceRange(String source, String destination, int offset, int upBound, int downBound, Date date) {
         Session session = SessionUtil.getSession();
         Transaction transaction = session.beginTransaction();
